test/requirements.txt
torch>=1.8.0
torchvision>=0.9.0
facenet-pytorch>=2.5.0
lpips>=0.1.3
Pillow>=8.0.0
scikit-image>=0.17.2
numpy>=1.19.2
